GUI/requirements.txt
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
